ImageLocations.txt created, tracks where we downloaded images.
Instead of making commits every time we download an image we created a
file that will track the locations of where we download the images.
Right now we're working on mockups so we feel comfortable downloading
almost any image we see.  With the final product we'll be more careful
with selecting images that we can actually use.